refactor(kaylendog/nodes): refactor peer tasks into modules
